Fix: 'and', 'or'. Move: LookAndSay test
